+++ ESFramework/Foundation/NSString+ESAdditions.h
@@ -12,6 +12,18 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface NSString (ESAdditions)
 
+- (nullable NSNumber *)numberValue;
+
+@property (readonly) char charValue;
+@property (readonly) unsigned char unsignedCharValue;
+@property (readonly) short shortValue;
+@property (readonly) unsigned short unsignedShortValue;
+@property (readonly) unsigned int unsignedIntValue;
+@property (readonly) long longValue;
+@property (readonly) unsigned long unsignedLongValue;
+@property (readonly) unsigned long long unsignedLongLongValue;
+@property (readonly) NSUInteger unsignedIntegerValue;
+
 - (NSString *)trimmedString;
 
 - (BOOL)contains:(NSString*)string;